tools/kb_enrichment/requirements.txt
# Google Cloud AI Platform
google-cloud-aiplatform>=1.38.0

# Vertex AI Generative AI
google-generativeai>=0.3.0

# Environment management
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Similarity computation
scikit-learn>=1.3.0
scipy>=1.11.0

# Progress tracking
tqdm>=4.66.0

# Existing project dependencies (for compatibility)
pydantic>=2.5.0
networkx>=3.2.0
